Improve performance of the sorting function
